pyproject.toml
[project]
name = "eacy"
version = "0.1.0"
description = "Everything-as-Code voice-driven knowledge and automation backbone"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "jsonschema>=4.23.0",
  "pydantic>=2.8.2",
  "PyYAML>=6.0.2",
  "typer>=0.12.5",
]

[dependency-groups]
dev = [
  "pytest>=8.3.2",
  "ruff>=0.6.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
